[Decomp][Chat] Cohesion split: extract help/command tables to ChatHelp.cpp
