+++ src/game/WorldHandlers/Chat.cpp
@@ -1651,133 +1651,7 @@ bool ChatHandler::ParseCommands(const char* text)
     return true;
 }
 
-/**
- * @brief Displays the available subcommands for a command table.
- *
- * @param table The subcommand table to display.
- * @param cmd The parent command name.
- * @return true if any subcommands were shown; otherwise false.
- */
-bool ChatHandler::ShowHelpForSubCommands(ChatCommand* table, char const* cmd)
-{
-    std::string list;
-    for (uint32 i = 0; table[i].Name != NULL; ++i)
-    {
-        // must be available (ignore handler existence for show command with possible available subcommands
-        if (!isAvailable(table[i]))
-        {
-            continue;
-        }
 
-        if (m_session)
-        {
-            list += "\n    ";
-        }
-        else
-        {
-            list += "\n\r    ";
-        }
-
-        list += table[i].Name;
-
-        if (table[i].ChildCommands)
-        {
-            list += " ...";
-        }
-    }
-
-    if (list.empty())
-    {
-        return false;
-    }
-
-    if (table == getCommandTable())
-    {
-        SendSysMessage(LANG_AVIABLE_CMD);
-        SendSysMessage(list.c_str());
-    }
-    else
-    {
-        PSendSysMessage(LANG_SUBCMDS_LIST, cmd);
-        SendSysMessage(list.c_str());
-    }
-    return true;
-}
-
-/**
- * @brief Displays help text and subcommands for a command path.
- *
- * @param table The command table to search.
- * @param cmd The command path to resolve.
- * @return true if help content was shown; otherwise false.
- */
-bool ChatHandler::ShowHelpForCommand(ChatCommand* table, const char* cmd)
-{
-    char const* oldCmd = cmd;
-    ChatCommand* command = NULL;
-    ChatCommand* parentCommand = NULL;
-
-    ChatCommand* showCommand = NULL;
-    ChatCommand* childCommands = NULL;
-
-    ChatCommandSearchResult res = FindCommand(table, cmd, command, &parentCommand);
-
-    switch (res)
-    {
-        case CHAT_COMMAND_OK:
-        {
-            // for "" subcommand use parent command if any for subcommands list output
-            if (strlen(command->Name) == 0 && parentCommand)
-            {
-                showCommand = parentCommand;
-                cmd = "";
-            }
-            else
-            {
-                showCommand = command;
-            }
-
-            childCommands = showCommand->ChildCommands;
-            break;
-        }
-        case CHAT_COMMAND_UNKNOWN_SUBCOMMAND:
-            showCommand = command;
-            childCommands = showCommand->ChildCommands;
-            break;
-        case CHAT_COMMAND_UNKNOWN:
-            // not show command list at error in first level command find fail
-            childCommands = table != getCommandTable() || strlen(oldCmd) == 0 ? table : NULL;
-            command = NULL;
-            break;
-    }
-
-    if (command && !command->Help.empty())
-    {
-        std::string helpText = command->Help;
-
-        // Attemp to localize help text if not in CLI mode
-        if (m_session)
-        {
-            int loc_idx = m_session->GetSessionDbLocaleIndex();
-            sCommandMgr.GetCommandHelpLocaleString(command->Id, loc_idx, &helpText);
-        }
-
-        SendSysMessage(helpText.c_str());
-    }
-
-    if (childCommands)
-        if (ShowHelpForSubCommands(childCommands, showCommand ? showCommand->Name : ""))
-        {
-            return true;
-        }
-
-    if (command && command->Help.empty())
-    {
-        SendSysMessage(LANG_NO_HELP_CMD);
-    }
-
-    return command || childCommands;
-}
 
 /**
  * @brief Validates chat text and embedded shift-links against configured strictness rules.
